requirements.txt
# Core ML & Retrieval
faiss-cpu==1.7.4
sentence-transformers==2.3.1
rank-bm25==0.2.2

# LLM & API
openai==1.3.0
tiktoken==0.5.1

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# UI & Visualization
streamlit==1.28.1
plotly==5.17.0

# Monitoring & Logging
prometheus-client==0.18.0
structlog==23.2.0

# Data Processing
pandas==2.1.3
pyarrow==14.0.1
numpy==1.24.3

# Utilities
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.0

